+++ src/configs/javascript.ts
@@ -68,7 +68,7 @@ export function javascript(options: JavaScriptOptions = {}): Linter.Config {
 			'no-class-assign': 'error',
 			'no-compare-neg-zero': 'error',
 			'no-cond-assign': ['error', 'always'],
-			'no-console': ['error', { allow: ['warn', 'error'] }],
+			'no-console': ['error', { allow: ['info', 'warn', 'error'] }],
 			'no-const-assign': 'error',
 			'no-control-regex': 'error',
 			'no-debugger': 'error',